refactor: refactored architecture
