+++ src/components/Common/PrivateRoute/PrivateRoute.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
 import { connect } from "react-redux";
-import authSelector from "./selectors/auth";
+import authSelector from "../../../selectors/auth";
 
 export const PrivateRoute = ({ component: Component, isLoggedIn, ...rest }) => (
   <Route
