+++ src/pages/Registration/Registration.css
@@ -1,5 +1,5 @@
 .registration {
-  background-image: url("./login-background.jpg");
+  background-image: url("../../assets/login-background.jpg");
   background-size: cover;
   height: 100vh;
 }
